chore: moved pixel rescaling to parent View
